pyproject.toml
[tool.poetry]
name = "emis-task"
version = "0.1.0"
description = ""
authors = ["Jay <jay.veer1@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
isort = "^5.13.2"
sqlalchemy = "^2.0.29"
pymysql = "^1.1.0"
python-dotenv = "^1.0.1"
pytest-mock = "^3.14.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
flake8 = "^7.0.0"
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
